requirements.txt
# Dependências para o modelo de Deep Learning
# TensorFlow e Keras
# Versão CPU para compatibilidade com sistemas sem GPU dedicada 
# Outras bibliotecas para manipulação de dados, visualização e conexão com banco de dados 
# Autor: Pedro Falcão
tensorflow-cpu==2.15.0
scikit-learn
h5py
numpy
pandas
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
matplotlib>=3.5.0
seaborn>=0.11.0
papermill>=2.4.0
jupyter>=1.0.0
psycopg2-binary
papermill 
nbformat